+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 2.13.0
+
+* Add support for Ecowitt Black Globe Temperature Sensor (WN38)
+* Add support for Ecowitt Mini Rain Gauge (WN20)
+* Implement soil conductivity characteristic for WH52
+* Display service for `last24hrain` data property if present in data report
+
 ### 2.12.0
 
 * Add support for Ecowitt WH52 (contrib from [@joshuaappleman](https://github.com/joshuaappleman))
@@ -65,6 +72,8 @@
 * Add Homebridge verified badge
 
 ### 2.4.1
+
+BREAKING: See [v2 migration guide](https://github.com/rhockenbury/homebridge-ecowitt-weather-sensors/wiki/V2-Migration-Guide)
 
 * Promote v2 to GA
 * Improve data report simulator with parameters set in definitions file